When grouping, open conversations and send messages in separate passes
The conversations_open() call has sometimes been failing with

    {'ok': False, 'error': 'fatal_error'}

which causes the pairings to be aborted partway through processing a
channel, and also causes the channel costs not to be updated.  Open all
the conversations first, so a failure there will leave the channel
completely ungrouped, and then send messages in a separate pass.